+++ src/partials/html/catalog.html
@@ -1,23 +1,4 @@
 <section>
-  <style>
-    .catalog-films {
-      display: flex;
-      flex-wrap: wrap;
-    }
-    .catalog-card {
-      width: calc(100% / 3);
-    }
-    .pagination {
-      display: flex;
-    }
-    .pagination-btn {
-      display: block;
-      padding: 2px 5px;
-    }
-    .btn-active > .pagination-btn {
-      background-color: red;
-    }
-  </style>
   <div class="container">
     <h2 class="visually-hidden">Search Form</h2>
     <form class="search-form">
@@ -38,6 +19,5 @@
       <h2 class="visually-hidden">Catalog Films</h2>
       <!-- Карточки изображений -->
     </div>
-    <div id="pagination"></div>
   </div>
 </section>
