+++ src/partials/js/catalog/createCards.js
@@ -4,9 +4,7 @@ export const createCards = dataCard => {
   const markup = dataCard
     .map(card => {
       return `<div data-id=${card.id} class="catalog-card">
-      <img src='https://image.tmdb.org/t/p/original${card.poster}' alt="${
-        card.title
-      }" loading="lazy" />
+      <img src='https://image.tmdb.org/t/p/original${card.poster}' alt="${card.title}" loading="lazy" />
       <div class="catalog-info">
         <div>
           <p class="catalog-info-name">
@@ -19,7 +17,7 @@ export const createCards = dataCard => {
         </div>
         <div
           class="Stars"
-          style="--rating: ${card.rating / 2}"
+          style="--rating: ${card.rating}"
           aria-label="Rating of this product is 2.3 out of 5."
         ></div>
       </div>
@@ -27,7 +25,7 @@ export const createCards = dataCard => {
     })
     .join('');
   addHTML(markup);
-  //console.log(markup);
+  console.log(markup);
   //   dataCard.map(el => console.log(el));
 };
